Clarify various concepts and tighten text
- Clarified a few concepts
- Tightened the text

The "Motivation Behind .NET Core" still needs help. It doesn't correctly describe the primary motivation for the project. As I said before, .NET Core wasn't built to solve the fragmentation issue, but to provide a cross-platform version of .NET. CoreFX was built to be componentized to enable smaller distributions. It was also built to solve the fragmentation issue, but as a secondary concern.