Update startup.rst, trying to make one sentence more readable. (#1911)
* Update startup.rst

I noticed that I had to read that sencence five times before coming to terms with it. I hope that my suggestion helps to make it more readable.

* Delete obsolete passage about DI

As @RickAnderson suggests, I just delete the passage that is trying to explain Dependency Injection in one sentence (and fails). He's right: all the info about Dependency Injection is available through the DI-link in the first sentence of that paragraph.